requirements.txt
ffmpeg-python==0.2.0
pyqt6==6.7.1
pytubefix==8.12.3